tests: move job-archive interface tests
Move the unit tests related to the job-archive
interface to their own test file
